requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.8.0.76
numpy==1.26.0
Pillow==10.0.0
python-dotenv==1.0.0
ultralytics==8.0.20
transformers==4.35.0
torch==2.1.0

